@@ -1,34 +1,34 @@
 {
-  "name": "spectrum-dreamscape",
-  "version": "1.0.0",
-  "description": "Shows how to form self-contained components with their own state and user interaction.",
-  "keywords": [
-    "interaction",
-    "pointer-events"
-  ],
-  "main": "src/index.js",
+  "name": "my-project",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@react-three/drei": "9.92.0",
     "@react-three/fiber": "8.15.12",
     "@types/three": "0.158.3",
     "aframe": "^1.6.0",
+    "autoprefixer": "^10.4.19",
     "networked-aframe": "^0.12.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-scripts": "5.0.1",
+    "postcss": "^8.4.40",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "tailwindcss": "^3.4.7"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "start:spec": "node ./server/easyrtc-server.js"
-  },
-  "browserslist": [
-    ">1%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "devDependencies": {
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.3",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.7",
+    "vite": "^5.3.4"
+  }
 }
